+++ src/core/sdl_wrapper.cpp
@@ -1,0 +1,144 @@
+/*
+ * SPDX-License-Identifier: BSD-3-Clause
+ * 
+ * Copyright (c) 2022, Saveliy Pototskiy (SavaLione) (savalione.com)
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. Neither the name of the copyright holder nor the names of its contributors
+ *    may be used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE REGENTS AND CONTRIBUTORS ``AS IS'' AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED.  IN NO EVENT SHALL THE REGENTS OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
+ * OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+ * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
+ * SUCH DAMAGE.
+ */
+/**
+ * @file
+ * @brief SDL2 wrapper
+ * @author Saveliy Pototskiy (SavaLione)
+ * @date 26 Nov 2022
+ */
+#include "core/sdl_wrapper.h"
+
+#include <spdlog/spdlog.h>
+
+sdl_wrapper::sdl_wrapper()
+{
+    /* SDL 2.0 initialization */
+    if(SDL_Init(SDL_INIT_VIDEO | SDL_INIT_TIMER | SDL_INIT_GAMECONTROLLER) != 0)
+    {
+        std::string _err = "Error initialization SDL2: ";
+        _err += SDL_GetError();
+        throw std::runtime_error(_err);
+    }
+
+    // GL 3.0 + GLSL 130
+    const char *glsl_version = "#version 130";
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_FLAGS, 0);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_CORE);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 3);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 0);
+
+    window = SDL_CreateWindow(_name.c_str(), 0, 0, 1024, 1024, SDL_WINDOW_OPENGL | SDL_WINDOW_SHOWN);
+    if(!window)
+    {
+        std::string _err = "Error creating SDL2 window: ";
+        _err += SDL_GetError();
+        throw std::runtime_error(_err);
+    }
+
+    context = SDL_GL_CreateContext(window);
+    if(!context)
+    {
+        std::string _err = "Error creating SDL2 context: ";
+        _err += SDL_GetError();
+        throw std::runtime_error(_err);
+    }
+
+    // Enable glew experimental, this enables some more OpenGL extensions.
+    glewExperimental = GL_TRUE;
+    if(glewInit() != GLEW_OK)
+    {
+        throw std::runtime_error("Failed to initialize GLEW");
+    }
+
+    /* SDL variables */
+    SDL_GetWindowSize(window, &window_width, &window_height);
+    spdlog::info("Window width: {} height: {}", window_width, window_height);
+
+    /* Some OpenGL settings */
+    glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
+    glClearDepth(1.0f);
+    glEnable(GL_DEPTH_TEST);
+}
+
+sdl_wrapper::~sdl_wrapper()
+{
+    /* SDL */
+    SDL_GL_DeleteContext(context);
+    SDL_DestroyWindow(window);
+    SDL_Quit();
+}
+
+void sdl_wrapper::pool_event()
+{
+    while(SDL_PollEvent(&event))
+    {
+        switch(event.type)
+        {
+            case SDL_KEYDOWN:
+                switch(event.key.keysym.sym)
+                {
+                    case SDLK_ESCAPE:
+                        _exit = true;
+                        break;
+                    default:
+                        break;
+                }
+                break;
+            case SDL_MOUSEBUTTONDOWN:
+                spdlog::info("Touch x: {} y: {}", event.button.x, event.button.y);
+                break;
+            case SDL_QUIT:
+                _exit = true;
+                break;
+            default:
+                break;
+        }
+    }
+}
+
+void sdl_wrapper::loop() {}
+
+void sdl_wrapper::run()
+{
+    /* Initialization */
+    init();
+
+    /* Main loop */
+    while(!_exit)
+    {
+        loop();
+        pool_event();
+        
+        SDL_GL_SwapWindow(window);
+    }
+}
+
+void sdl_wrapper::init() {}