+++ src/core/sdl_wrapper.h
@@ -1,0 +1,73 @@
+/*
+ * SPDX-License-Identifier: BSD-3-Clause
+ * 
+ * Copyright (c) 2022, Saveliy Pototskiy (SavaLione) (savalione.com)
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. Neither the name of the copyright holder nor the names of its contributors
+ *    may be used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE REGENTS AND CONTRIBUTORS ``AS IS'' AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED.  IN NO EVENT SHALL THE REGENTS OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
+ * OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+ * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
+ * SUCH DAMAGE.
+ */
+/**
+ * @file
+ * @brief SDL2 wrapper
+ * @author Saveliy Pototskiy (SavaLione)
+ * @date 26 Nov 2022
+ */
+#ifndef CORE_SDL_WRAPPER_H
+#define CORE_SDL_WRAPPER_H
+
+// clang-format off
+#include <CL/cl_gl.h>
+#include <GL/glew.h>
+#include <SDL2/SDL.h>
+#include <SDL2/SDL_opengl.h>
+#include <GL/gl.h>
+// clang-format on
+
+#include <string>
+
+class sdl_wrapper
+{
+public:
+    sdl_wrapper();
+    ~sdl_wrapper();
+    void run();
+
+protected:
+    virtual void pool_event();
+    virtual void loop();
+    virtual void init();
+
+    /* SDL */
+    SDL_Window *window;
+    SDL_GLContext context;
+    SDL_Event event;
+
+    int window_width  = 0;
+    int window_height = 0;
+    bool _exit        = false;
+    std::string _name = "nyx";
+};
+
+#endif // CORE_SDL_WRAPPER_H